refactor: extract AssistantFormHeader from AssistantForm (SRP)
