+++ frontend/svelte-app/src/lib/components/assistants/AssistantForm.svelte
@@ -18,6 +18,7 @@
 	import TemplateSelectModal from '$lib/components/modals/TemplateSelectModal.svelte'; // Import template modal
 	import { openTemplateSelectModal } from '$lib/stores/templateStore'; // Import template store function
 	import { sanitizeName } from '$lib/utils/nameSanitizer'; // Import sanitization utility
+import AssistantFormHeader from './AssistantFormHeader.svelte';
 	import { getAssistantMetadataObject } from '$lib/utils/assistantData';
 
 	const dispatch = createEventDispatcher(); // For dispatching success event
@@ -1157,13 +1158,6 @@
 	// --- Import Functionality ---
 
 	/**
-	 * Triggers the hidden file input click.
-	 */
-	function triggerFileInput() {
-		document.getElementById('import-assistant-json')?.click();
-	}
-
-	/**
 	 * Handles file selection for import.
 	 * @param {Event} event
 	 */
@@ -1390,54 +1384,13 @@
 </script>
 
 	<div class="p-4 md:p-6 border rounded-md shadow-sm bg-white">
-	<!-- Header Section (Title and Top Buttons) -->
-	<div class="mb-6 pb-4 border-b border-gray-200">
-		<div class="flex justify-between items-center">
-			<h2 class="text-2xl font-semibold text-brand">
-				{#if formState === 'create'}
-					{$_('assistants.form.titleCreate', { default: 'Create New Assistant' })}
-				{:else}
-					{$_('assistants.form.titleViewEdit', { default: 'Assistant Details' })}
-					{#if initialAssistantData?.id} (ID: {initialAssistantData.id}){/if}
-				{/if}
-			</h2>
-			
-			<!-- Import Button - Only show in create mode -->
-			{#if formState === 'create'}
-				<div>
-					<button 
-						type="button" 
-						onclick={triggerFileInput}
-						class="inline-flex items-center px-4 py-2 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-brand"
-					>
-						<svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-							<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M9 19l3 3m0 0l3-3m-3 3V10"></path>
-						</svg>
-						{$_('assistants.form.import.button', { default: 'Import from JSON' })}
-					</button>
-				</div>
-			{/if}
-		</div>
-		
-		<!-- Import Error Message -->
-		{#if importError}
-			<div class="mt-3 p-3 border border-red-200 bg-red-50 rounded-md">
-				<div class="flex">
-					<div class="flex-shrink-0">
-						<svg class="h-5 w-5 text-red-400" fill="currentColor" viewBox="0 0 20 20">
-							<path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clip-rule="evenodd"></path>
-						</svg>
-					</div>
-					<div class="ml-3">
-						<h3 class="text-sm font-medium text-red-800">
-							{$_('assistants.form.import.error', { default: 'Import Error' })}
-						</h3>
-						<p class="mt-1 text-sm text-red-700">{importError}</p>
-					</div>
-				</div>
-			</div>
-		{/if}
-	</div>
+
+	<AssistantFormHeader
+		{formState}
+		assistantId={initialAssistantData?.id}
+		{importError}
+		onImportFile={handleFileSelect}
+	/>
 
 	{#if $assistantConfigStore.loading && !configInitialized}
 		<p class="text-center text-gray-600 py-10">{$_('assistants.loadingConfig', { default: 'Loading configuration...' })}</p>
@@ -2100,15 +2053,6 @@
 
 		</form>
 	{/if} 
-
-	<!-- Hidden file input for import functionality -->
-	<input 
-		type="file" 
-		id="import-assistant-json" 
-		accept=".json" 
-		onchange={handleFileSelect}
-		style="display: none;"
-	/>
 
 	<!-- Template Selection Modal -->
 	<TemplateSelectModal onSelect={handleTemplateSelected} />
